+++ CHANGELOG.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+2.0.0 (2024-11-06)
+------------------
+After the automated 'Chemicals delivered' email has been send to the user,
+revert all buttons back to their idle off state. The user no longer has to
+deactivate any lit buttons, and hence the 'All done' email message is now
+suppressed.
+
 1.0.0 (2021-10-28)
 ------------------
 * Initial release